Remove temp maple2
